Extract OfficeFormatter from PastedContentFormatter
The Word/Outlook list rebuilding was nine private methods deep in a class
that otherwise does small, unrelated paste repairs. Move it into its own
collaborator: PastedContentFormatter instantiates OfficeFormatter and sends
it .format().

The generic repairs stay put. #stripStyleElements, #unwrapPlaceholderAnchors
and #stripTableCellColorStyles apply to all pasted content, and the stray-list
repairs (#unwrapWrappedListChildren, #nestStrayListChildren,
#stripStrayListChildren) start from querySelectorAll("ol, ul"), so they run for
Gmail and rendered-view pastes too, not just Office. They have to keep running
after the rebuild, which is why the instantiation sits mid-pipeline.

Private names drop their now-redundant Office prefix (#officeListLevel becomes
#listLevel) since the class carries it. Bodies are otherwise unchanged.
