+++ src/editor/contents/pasted_content_formatter.js
@@ -1,3 +1,5 @@
+import OfficeFormatter from "./pasted_content_formatter/office_formatter"
+
 export default class PastedContentFormatter {
   constructor(doc) {
     this.doc = doc
@@ -7,7 +9,7 @@ export default class PastedContentFormatter {
     this.#stripStyleElements()
     this.#unwrapPlaceholderAnchors()
     this.#stripTableCellColorStyles()
-    this.#rebuildOfficeLists()
+    new OfficeFormatter(this.doc).format()
     this.#unwrapWrappedListChildren()
     this.#nestStrayListChildren()
     this.#stripStrayListChildren()
@@ -48,143 +50,6 @@ export default class PastedContentFormatter {
       cell.style.removeProperty("background")
       cell.style.removeProperty("color")
     }
-  }
-
-  // Word and Outlook never emit <ul>/<ol>/<li>. They simulate lists with flat
-  // paragraphs that declare their depth in an inline `mso-list:l0 level2 lfo1`
-  // and carry the bullet or number glyph in a <span style="mso-list:Ignore">.
-  // Nothing downstream understands either, so the items land as plain paragraphs
-  // with the marker baked into the text. Rebuild real lists from the level each
-  // paragraph declares.
-  //
-  // Everything this reads lives in inline attributes, so it stays correct after
-  // #stripStyleElements has dropped Office's style sheet.
-  #rebuildOfficeLists() {
-    for (const run of this.#officeListRuns()) {
-      this.#replaceWithLists(run)
-    }
-  }
-
-  // A run is a stretch of Office list paragraphs that are siblings with nothing
-  // but whitespace and Word's conditional comments between them. Anything else
-  // in between (a heading, a plain paragraph) ends one list and starts another.
-  #officeListRuns() {
-    const runs = []
-    const claimed = new Set()
-
-    for (const paragraph of this.doc.querySelectorAll("p[style*='mso-list']")) {
-      if (claimed.has(paragraph) || this.#officeListLevel(paragraph) === null) continue
-
-      const run = []
-      let node = paragraph
-      while (node) {
-        run.push(node)
-        claimed.add(node)
-        node = this.#nextOfficeListParagraph(node)
-      }
-      runs.push(run)
-    }
-
-    return runs
-  }
-
-  #nextOfficeListParagraph(paragraph) {
-    let sibling = paragraph.nextSibling
-    while (sibling && this.#isIgnorableBetweenOfficeListParagraphs(sibling)) {
-      sibling = sibling.nextSibling
-    }
-
-    if (sibling && sibling.nodeType === Node.ELEMENT_NODE && this.#officeListLevel(sibling) !== null) {
-      return sibling
-    }
-
-    return null
-  }
-
-  #isIgnorableBetweenOfficeListParagraphs(node) {
-    if (node.nodeType === Node.COMMENT_NODE) {
-      return true
-    }
-    return node.nodeType === Node.TEXT_NODE && node.textContent.trim() === ""
-  }
-
-  // Word numbers levels from 1, but a copied selection often starts partway down
-  // a list, so every paragraph declares level3 with no level1 above it. Anchoring
-  // the run at its own shallowest level is what keeps the result from arriving
-  // needlessly deep.
-  #replaceWithLists(paragraphs) {
-    const levels = paragraphs.map((paragraph) => this.#officeListLevel(paragraph))
-    const baseLevel = Math.min(...levels)
-    const roots = []
-    const stack = []
-
-    paragraphs.forEach((paragraph, index) => {
-      const depth = levels[index] - baseLevel + 1
-      const tagName = this.#officeListTagName(paragraph)
-
-      while (stack.length > depth) {
-        stack.pop()
-      }
-      if (stack.length === depth && stack[depth - 1].tagName !== tagName.toUpperCase()) {
-        stack.pop()
-      }
-
-      while (stack.length < depth) {
-        const list = this.doc.createElement(tagName)
-        if (stack.length === 0) {
-          roots.push(list)
-        } else {
-          this.#lastItemOf(stack[stack.length - 1]).appendChild(list)
-        }
-        stack.push(list)
-      }
-
-      stack[stack.length - 1].appendChild(this.#createOfficeListItem(paragraph))
-    })
-
-    paragraphs[0].replaceWith(...roots)
-    for (const paragraph of paragraphs.slice(1)) {
-      paragraph.remove()
-    }
-  }
-
-  #lastItemOf(list) {
-    if (!list.lastElementChild) {
-      list.appendChild(this.doc.createElement("li"))
-    }
-    return list.lastElementChild
-  }
-
-  // Word writes bullets as the raw glyph of the marker font (Symbol's "·",
-  // Wingdings' "§", Courier New's "o") and numbers as the rendered counter
-  // ("1.", "a)", "iv)"). A bare letter with no delimiter after it is Courier
-  // New's bullet, not a counter, so the delimiter is what tells them apart.
-  #officeListTagName(paragraph) {
-    const marker = paragraph.querySelector("[style*='mso-list:Ignore']")
-    const text = marker?.textContent || ""
-
-    if (/\d/.test(text) || /^\(?[A-Za-z]+[.)]/.test(text)) {
-      return "ol"
-    }
-    return "ul"
-  }
-
-  #createOfficeListItem(paragraph) {
-    for (const marker of paragraph.querySelectorAll("[style*='mso-list:Ignore']")) {
-      marker.remove()
-    }
-
-    const item = this.doc.createElement("li")
-    item.append(...paragraph.childNodes)
-    return item
-  }
-
-  #officeListLevel(element) {
-    const match = /mso-list:[^;"']*\blevel(\d+)/.exec(element.getAttribute("style") || "")
-    if (match) {
-      return parseInt(match[1], 10)
-    }
-    return null
   }
 
   // Some sources wrap runs of <li>s in a stray element (e.g. a <div> directly
